Bug fix: (Deleting message - related to messages in the chat and chat list) + Clean Up
Everything works as expected when the owner of the message tries to delete it. However, if there is a request from another user to delete a message sent by the sender (a message that the requester did not send), it wouldn't be deleted from the owner's side. This issue has now been fixed.